MANTA-30 cleanup warnOnce
Note using a non-const static like this won't scale as a library
function -- what happens if another unrelated class were to call
this before the vcf parser? Even in it's new form, this is
dodgey (ie. it's not thread safe) consider a generic 'exception
or warn' object to generalize the case in a safe way.