MANTA-253 control order of I<->D transitions
This edit allows insert to deletion changes in the intron
alingment to occur in only one order. It shouldn't change
output except to force the ID vs DI order when a combined
insert/delete variant occurs. This doesn't improve anything
except to reduce the size of the plausible path space. It
also standardizes the intron aligner with the other aligners
wrt this detail.