MANTA-257 change isect node evidence count scheme
This changes the intersecting node evidence sum to go from
using the 'leftmost' node only to using the max value of the
(left,right) nodes. Both new and old schemes are used in this
situation because the actual sum would artificially
double the merged node's evidence.

This should fix MANTA-257