MANTA-267 Support for contig names with colons.
This change is motivated by HLA contigs in the 1kg flavor of
hg38. To support this change all region parsing has been
unified to the htsapi library in manta (and Manta's python region
parser). We aren't using the htslib parser anymore.

The new behavior is to split on the last colon. If a colon is found,
then evaluate the second field as a begin-end position pair. If the
string can't be parsed this way then assume the region following the
last colon is also part of the contig name.

Note this means that the samtools "contig:begin" format without listing
an end position is no longer supported.