MANTA-1293 Add chrom name for repeated read error
A somewhat frequent error occurs when two primaty bam records have the
same QNAME and read_no. This is not only disallowed in the spec but is
an important QC safeguard against accidentally double-counting evidence.

The error message has been improved to replace tid with the chrom name
and to use the 1-index position number for each fo the two alignment
records reported in the read name collision.